rename files and types.
