+++ index.d.ts
@@ -1,5 +1,5 @@
 export declare class Extract {
-    public static run(filePath: string): Promise<IContent>;
+    public static run(filePath: string): Promise<IPackageMetadata>;
 }
 
 export declare class ExtractError extends Error {
@@ -19,7 +19,7 @@ export interface IProvisioningProfile {
     pathName: string;
 }
 
-export interface IContent {
+export interface IPackageMetadata {
     originalFileName: string;
     displayName: string;
     name: string;
@@ -38,7 +38,7 @@ export interface IContent {
     hasProvisioning: boolean;
 }
 
-export interface IIpaContent extends IContent {
+export interface IIPAMetadata extends IPackageMetadata {
     provision: IProvisioningProfile;
     appexProvisioningProfiles: IProvisioningProfile[];
 }
