+++ src/apkContent.ts
